+++ Runtime/Loggers/FileLoggerAsset.cs
@@ -52,29 +52,14 @@ namespace DeNA.Anjin.Loggers
                     return _logger;
                 }
 
-                var args = new Arguments();
-                string path;
-                if (args.FileLoggerOutputPath.IsCaptured())
+                var path = PathUtils.GetOutputPath(this.outputPath, new Arguments().FileLoggerOutputPath);
+                if (string.IsNullOrEmpty(path))
                 {
-                    path = args.FileLoggerOutputPath.Value();
-                }
-                else if (!string.IsNullOrEmpty(outputPath))
-                {
-                    // ReSharper disable once PossibleNullReferenceException
-                    var outputRootPath = AutopilotState.Instance.settings.OutputRootPath;
-                    path = PathUtils.GetAbsolutePath(outputPath, outputRootPath);
-                }
-                else
-                {
-                    throw new InvalidOperationException("outputPath is not set.");
+                    Debug.LogWarning("File Logger output path is not set.");
+                    return null;
                 }
 
-                var directory = Path.GetDirectoryName(path);
-                if (!string.IsNullOrEmpty(directory) && !Directory.Exists(directory))
-                {
-                    Directory.CreateDirectory(directory);
-                }
-
+                _handler?.Dispose();
                 _handler = new FileLogHandler(path, timestamp);
                 _logger = new Logger(_handler) { filterLogType = filterLogType };
                 return _logger;
@@ -104,7 +89,7 @@ namespace DeNA.Anjin.Loggers
             }
         }
 
-        private class FileLogHandler : ILogHandler, IDisposable
+        private sealed class FileLogHandler : ILogHandler, IDisposable
         {
             private readonly StreamWriter _writer;
             private readonly bool _timestamp;
