+++ Runtime/Reporters/JUnitXmlReporter.cs
@@ -2,7 +2,6 @@
 // This software is released under the MIT License.
 
 using System;
-using System.IO;
 using System.Threading;
 using System.Xml.Linq;
 using Cysharp.Threading.Tasks;
@@ -48,17 +47,11 @@ namespace DeNA.Anjin.Reporters
                 throw new InvalidOperationException("Autopilot is not running");
             }
 
-            var path = GetOutputPath(this.outputPath, new Arguments());
+            var path = PathUtils.GetOutputPath(this.outputPath, new Arguments().JUnitReportPath);
             if (string.IsNullOrEmpty(path))
             {
-                Debug.LogWarning("JUnit XML report output path is not set.");
+                Debug.LogWarning("JUnit XML Reporter output path is not set.");
                 return;
-            }
-
-            var directory = Path.GetDirectoryName(path);
-            if (!string.IsNullOrEmpty(directory) && !Directory.Exists(directory))
-            {
-                Directory.CreateDirectory(directory);
             }
 
             var testCaseName = settings.Name;
@@ -69,32 +62,6 @@ namespace DeNA.Anjin.Reporters
             new XDocument(new XDeclaration("1.0", "utf-8", null), testSuites).Save(path);
 
             await UniTask.CompletedTask;
-        }
-
-        internal static string GetOutputPath(string outputPathField, Arguments args = null)
-        {
-            if (args == null)
-            {
-                args = new Arguments();
-            }
-
-            string path;
-            if (args.JUnitReportPath.IsCaptured())
-            {
-                path = args.JUnitReportPath.Value();
-            }
-            else if (!string.IsNullOrEmpty(outputPathField))
-            {
-                // ReSharper disable once PossibleNullReferenceException
-                var outputRootPath = AutopilotState.Instance.settings.OutputRootPath;
-                path = PathUtils.GetAbsolutePath(outputPathField, outputRootPath);
-            }
-            else
-            {
-                return null;
-            }
-
-            return path;
         }
 
         internal static XElement CreateTestCase(string name, float time, string message, string stackTrace,
